backend/requirements.txt
fastapi
uvicorn
pydantic
pymupdf
chromadb
langchain
langchain-google-genai
python-dotenv
langchain-chroma
langchain-text-splitters
langchain-huggingface
sentence-transformers
